% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaner.R
\name{dataCleaner}
\alias{dataCleaner}
\title{dataCleaner}
\usage{
dataCleaner(dat, nrecords, provided)
}
\arguments{
\item{dat}{A a data frame containing the data}

\item{nrecords}{An integer for threshold of minimum data observaations}

\item{provided}{A logical indicating if data was provided or downloaded}
}
\value{
the cleaned data
}
\description{
cleans precipitation and temperature data from Modularity Homework to easier to use format
}
\examples{
dataCleaner(precipdata,45,TRUE)
dataCleaner(tempdata, 40, FALSE)

}
