% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltevrn.R
\docType{package}
\name{ltevrn}
\alias{ltevrn}
\alias{ltevrn-package}
\title{ltevrn}
\description{
A package to organize functions for my EVRN 720 homework from the spring of 2018.
}
\section{ltevrn functions}{

add, subtract, dataCleaner, tempConvert, treeHeight, hello.R
}

