% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtract}
\alias{subtract}
\title{Subtract two numbers.}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A real number}

\item{y}{A real number}
}
\value{
the subtraction of \code{x} and \code{y}
}
\description{
Subtract two numbers.
}
\examples{
subtract(1,1)
subtract(10,2)
}
