% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempConvert.R
\name{tempConvert}
\alias{tempConvert}
\title{tempConvert}
\usage{
tempConvert(tempIn, inputflag)
}
\arguments{
\item{tempIn}{A real number or vector of real numbers}

\item{inputflag}{A string identifying units of input value(s)}
}
\value{
the conversion of \code{tempIn} as a real number or vector of real numbers
}
\description{
Convert temperature between Celcius and Fahrenheit
}
\examples{
tempConvert(65,"fahrenheit")
tempConvert(c(10,20,40), "celcius")

}
