% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeHeight.R
\name{treeHeight}
\alias{treeHeight}
\title{treeHeight}
\usage{
treeHeight(distance, theta, thetaflag)
}
\arguments{
\item{distance}{A real number or vector of real numbers}

\item{theta}{A real number or vector of real numbers}

\item{thetaflag}{A string identifying units of theta value(s)}
}
\value{
the height of the tree given \code{distance} and \code{theta}
}
\description{
Calculate tree height given distance away from tree and angle of observer
}
\examples{
treeHeight(100,45,"degrees")
treeHeight(c(10,20,40),c(0.785,0.524,0.785), "radians")

}
